Refactor CLI registration
