--- src/fairseq2/setup/_cli/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from __future__ import annotations
-
-from fairseq2.setup._cli._commands import setup_cli as setup_cli
